Add sync migration

+++ tests/utils.py
@@ -5,4 +5,5 @@ from app import collections
 
 def restart_db(db: StandardDatabase):
     for collection in collections.CollectionBase.__subclasses__():
-        db.collection(collection._collection_name).truncate()
+        if collection._collection_name:
+            db.collection(collection._collection_name).truncate()
